Fix indentation that incorrectly implies a different control flow.
The actual control flow is intentional while the indenting is incorrect.
This is intended to be a cosmetic change.

Verified that each of these was part of a KNF commit that wasn't intending
to change behavior.  Also, double checked against the history of changes in
OpenSSL and BoringSSL.

Addresses Coverity CIDs: 78842, 78859, 78863.

ok tedu@