libtls new feature: disable renegotation
It is a historically insecure feature that's increasingly obsolete.
It will be removed in TLSv1.3.

Renegotation is enabled by default even when app is unaware of it.
Why is this good?  When apps really need it they need to enable
it explicitly.

It's quite non-trivial to disable :(  This approach was done by
examining https://github.com/reyk/relayd repo.