libtls fix: secure "compat" ciphersuite
In short: "compat" suite enables EXP:LOW:MEDIUM ciphers
which is security problem on server-side.

The "default"/"secure" shortcut is very aggressive by it's
TLSv1.2+AEAD setting.  Which is cool, but that encourages
users to switch away from it when it's simply wants to support
slightly older TLSv1.0 clients.

The "compat" shortcut should be for those users wanting normal
compatibility with TLSv1 clients, not for those few that really
need DES.

Update: Because there have been clients with broken 3DES but
working RC4, enable MEDIUM and move 3DES to last.