Link to Peter Gutmann's classic "X.509 Style Guide". Thanks to otto@ for making me aware of it.
If people know newer documents that are similarly readable and
interesting, please speak up.  I hate sending people to the STANDARDS
only for more information.  On the one hand, that's torture, and
on the other hand, if i read Gutmann correctly, the standards
sometimes provide bad advice, and often none at all.