Make X509_check_host and X509_check_email calculate strlen if chklen is 0.
According to the X509_check_host(3) (src/lib/libcrypto/man/X509_check_host.3),
"The namelen argument must be the number of characters in the name string or zero,
in which case the length is calculated with strlen(name)" for both functions
mentioned above. Even though I think this not a good API (I would always make
the called pass the exact length and call it a day), this causes compatibility
problems with code written for OpenSSL, a good example is MariaDB's SSL certificate
hostname check found here:
https://github.com/MariaDB/server/blob/10.3/sql-common/client.c#L1825

This change uses the code from OpenSSL's, which on top of calculating the
string length also tolerates the caller including terminating NUL in string
length. This way, LibreSSL's code should be API compatible with OpenSSL's
interface. It would be good to add unit tests for these functions.