Avoid potential double-frees following EVP_CIPHER_CTX_copy().
In the case of a cipher with a custom copy control, if that control fails
we may still have pointers that we do not own in the previously copied
cipher data. Avoid potential double-frees by zeroing and freeing the
copied cipher data in this case.

Issue reported by Guido Vranken.

ok tb@