Refactor the server hello processing code in the TLSv1.3 client.
Use flags to signal the need to switch to the legacy client and to identify
a hello retry request. This allows the caller to take appropriate action,
rather than trying to do this in the parsing/processing code.

Split the key deriviation and record protection engagement code into a
separate function, both for readability and reuse.

Change handshake states outside of the processing code.

ok tb@