cstyle in illumos noticed some weird syntax, which this fixes.
some errx lines in if statements were terminated with commas, which
caused the following statement to be considered part of the error
handling. while it is bad style, ingo points out it was also a bug
which which caused some tests in the code to be skipped.

this reminds me of a haiku that Chris Pascoe (cpascoe@) had behind
his desk:

	Also, that comma
	Should be a semi-colon.
	Cherry blossoms fall.

this was found by Robert Mustacchi when porting the tests to illumos.
ok schwarze@ stsp@

thank you robert.