Remove remaining error *_str_functs[]
A number of years ago we dropped the concept of having function names in
errors, since it is not that useful and very quickly gets out of sync when
refactoring. It would seem that some new ones got imported and some missed
the last clean up.

ok tb@ beck@ "kill it with fire"