Add P-521 to the list of curves supported by default in the client.
A certain VPN provider appears to have configured their servers to only
accept P-521 for TLSv1.3 key exchange. The particular VPN software in use
also does not currently allow for the TLSv1.3 key share groups to be
configured, which means that there is no way to easily use LibreSSL in
this situation.

Include P-521 in the list of curves that are supported by default in the
client, in order to increase interoperability.

Discussed at length with beck@, inoguchi@ and tb@.

ok tb@