Split session retrieval out of ssl_get_prev_session()
In case the session ticket was empty or missing, an attempt is made to
retrieve the session from the internal cache or via a callback. This
code can easily be flattened a bit and factored into two functions. I
decided to wrap those into a third function to make the call from the
switch easier on the eye.

I could have kept the try_session_cache flag, but it now seems rather
pointless and awkwardly named anyway, so I took its negation and named
it ticket_decrypted.

To top things off, a little bit of polish in the exit path.

ok beck inoguchi jsing (with the usual healthy dose of nits)