Remove unnecessary zeroing after recallocarray(3)
Zap a memset that was redundant since OpenSSL 0.97b was merged by
markus in 2003. Nowadays it's otto's recallocarray(3) that does the
zeroing.

ok beck inoguchi otto

PS: ASN1_BIT_STRING_set_bit(3) was committed on Dec 21 1998 by Ralf S.
Engelschnall and used this bizarre allocation idiom:

                if (a->data == NULL)
                        c=(unsigned char *)Malloc(w+1);
                else
                        c=(unsigned char *)Realloc(a->data,w+1);

People complained about Malloc, Realloc and Free being used elsewhere, so
on Jun 1 2000, Richarde Levitte swept the OpenSSL tree and it became this.

                if (a->data == NULL)
                        c=(unsigned char *)OPENSSL_malloc(w+1);
                else
                        c=(unsigned char *)OPENSSL_realloc(a->data,w+1);

Then it was found that existing data should be cleaned, and on Nov 13 2002
Ben Laurie changed the last line to
                        c=(unsigned char *)OPENSSL_realloc_clean(a->data,
                                                                a->length,
                                                                w+1);