Refactor IPAddressFamily accessors
Introduce a helper function that allows fetching the AFI and the
optional SAFI out of an IPAddressFamily. Also add two wrappers that
only fetch and validate the AFI, where validation currently only
means that the length is between 2 and 3.

Use these accessors throughout to simplify and streamline the code.

ok inoguchi jsing