Document EVP_{CIPHER,MD}_do_all{,_sorted}(3)
The function prototypes in the SYNOPSIS don't look great, but schwarze
assures me that this is how it is supposed to be. It is rather strange
that OpenSSL chose to sprinkle OPENSSL_init_crypto() calls into these
four functions rather than two inside OBJ_NAME_do_all{,_sorted}(3).
Surely there was a good reason for that.

With input and fixes from schwarze