x86_64cpuid.pl: zap _CET_ENDBR opcode from the init section
Before this patch, an `endbr*` opcode was present in the `.init`
section.

This caused a crash on startup in MinGW builds for example (possibly
also in MSVC builds), as seen in 3.8.3 and 3.9.0 releases.

x86_64-xlate.pl is not prepared to handle CPU opcodes in this segment,
reinforced by its header comment:
  9. .init segment is allowed to contain calls to functions only.

Affected platforms expect a list of function pointers in the `.init`
section, and were crashing when making the first call, which was in
fact the `_CET_ENDBR` opcode, not a valid function address.

Ref: https://github.com/libressl/portable/issues/1015